refactor(volume-mask): address PR review feedback
- Add Tadej Satler to package.json contributors
- Add pointInPolygon2D to mol-math/geometry/polygon.ts; remove duplicate internal/polygon.ts
- Move MRC writer to mol-io/writer/ccp4/ccp4.ts (CCP4Writer namespace); mrc-export.ts becomes a thin browser download wrapper
- Split extension (headless compute) from UI: move VolumeMaskController + React components to src/examples/volume-mask/; extension now exports only VolumeMaskBehavior, MaskVolumeFromSource, and types
