+++ package.json
@@ -126,7 +126,8 @@
     "Victoria Doshchenko <doshchenko.victoria@gmail.com>",
     "Diego del Alamo <diego.delalamo@gmail.com>",
     "Tianzhen Lin (Tangent) <tangent@usa.net>",
-    "Russ Taylor <russ@reliasolve.com>"
+    "Russ Taylor <russ@reliasolve.com>",
+    "Tadej Satler <tadej.satler@gmail.com>"
   ],
   "license": "MIT",
   "devDependencies": {
