--- src/extensions/volume-mask/internal/polygon.ts
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2026 mol* contributors, licensed under MIT, See LICENSE file for more info.
- * @author Tadej Satler <tadej.satler@gmail.com>
- */
-
-export type Point2D = [number, number];
-
-/** Ray-casting point-in-polygon test. Handles edge cases at polygon vertices. */
-export function pointInPolygon(px: number, py: number, polygon: Point2D[]): boolean {
-    const n = polygon.length;
-    if (n < 3) return false;
-    let inside = false;
-    for (let i = 0, j = n - 1; i < n; j = i++) {
-        const xi = polygon[i][0], yi = polygon[i][1];
-        const xj = polygon[j][0], yj = polygon[j][1];
-        if (((yi > py) !== (yj > py)) && (px < ((xj - xi) * (py - yi)) / (yj - yi) + xi)) {
-            inside = !inside;
-        }
-    }
-    return inside;
-}
