+++ src/mol-math/geometry/_spec/polygon.spec.ts
@@ -5,7 +5,7 @@
  */
 
 import { Vec2 } from '../../linear-algebra';
-import { pointInPolygon } from '../polygon';
+import { pointInPolygon, pointInPolygon2D } from '../polygon';
 
 describe('pointInPolygon', () => {
     it('basic', () => {
@@ -17,5 +17,27 @@ describe('pointInPolygon', () => {
         ];
         expect(pointInPolygon(Vec2.create(0, 0), polygon, 4)).toBe(true);
         expect(pointInPolygon(Vec2.create(2, 2), polygon, 4)).toBe(false);
+    });
+});
+
+describe('pointInPolygon2D', () => {
+    const square: [number, number][] = [[-1, -1], [1, -1], [1, 1], [-1, 1]];
+
+    it('inside square', () => {
+        expect(pointInPolygon2D(0, 0, square)).toBe(true);
+    });
+
+    it('outside square', () => {
+        expect(pointInPolygon2D(2, 2, square)).toBe(false);
+    });
+
+    it('degenerate polygon (< 3 vertices)', () => {
+        expect(pointInPolygon2D(0, 0, [[0, 0], [1, 0]])).toBe(false);
+    });
+
+    it('triangle', () => {
+        const tri: [number, number][] = [[0, 0], [1, 0], [0.5, 1]];
+        expect(pointInPolygon2D(0.5, 0.4, tri)).toBe(true);
+        expect(pointInPolygon2D(0.5, 1.1, tri)).toBe(false);
     });
 });
