+++ src/mol-math/geometry/polygon.ts
@@ -22,3 +22,17 @@ export function pointInPolygon(point: Vec2, polygon: NumberArray, count: number)
     }
     return inside;
 }
+
+/** raycast along x-axis and apply even-odd rule; accepts an array of [x, y] tuples */
+export function pointInPolygon2D(px: number, py: number, polygon: readonly [number, number][]): boolean {
+    const n = polygon.length;
+    if (n < 3) return false;
+    let inside = false;
+    for (let i = 0, j = n - 1; i < n; j = i++) {
+        const xi = polygon[i][0], yi = polygon[i][1];
+        const xj = polygon[j][0], yj = polygon[j][1];
+        if (((yi > py) !== (yj > py)) && (px < ((xj - xi) * (py - yi)) / (yj - yi) + xi))
+            inside = !inside;
+    }
+    return inside;
+}
